server/requirements.txt
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jieba==0.42.1
json2html==1.3.0
lxml==4.5.1
parsel==1.6.0
Pillow==7.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
